requirements.txt
numpy >= 1.17.0
easydict >= 1.9
scipy >= 1.5.2
matplotlib >= 3.1.3
Pillow >= 6.2.0
mindspore == 1.2.0
wheel >= 0.32.0
setuptools >= 40.8.0
requests >= 2.22.0
flask >= 1.1.1
python-Levenshtein >= 0.10.2
gensim >= 3.8.1
PyYAML
pycocotools >= 2.0.0          # for st test